completed refactoring so QUINTS can be set to either long or bigint

+++ primes.cc
@@ -240,7 +240,7 @@ Quad Factorization::solve_CRT(const vector<Quad>& v) // solution to x=v[i] mod Q
 
 void Quadprimes::init(long maxn)
 {
-  maxnorm = BIGINT(maxn);
+  maxnorm = maxn;
   vector<Quadprime> list1, list2;
 
   // First fill up lists of degree 1 and degree 2 primes
